ExceptionHandler added

+++ src/main/java/com/in28minutes/springboot/web/controller/TodoController.java
@@ -50,6 +50,8 @@ public class TodoController {
 
     @RequestMapping(value = "/delete-todo", method = RequestMethod.GET)
     public String deleteTodo(@RequestParam int id) {
+        if(id == 1)
+            return "error";
         service.deleteTodo(id);
         return "redirect:/list-todos";
     }
